requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
bzip2=1.0.8=h99b78c6_7
ca-certificates=2024.7.4=hf0a4a13_0
dnspython=2.6.1=pypi_0
kafka-python=2.0.2=pypi_0
libffi=3.4.2=h3422bc3_5
libsqlite=3.46.0=hfb93653_0
libzlib=1.3.1=hfb2fe0b_1
ncurses=6.5=hb89a1cb_0
numpy=2.0.1=pypi_0
opencv-python=4.10.0.84=pypi_0
openssl=3.3.1=hfb2fe0b_2
pandas=2.2.2=pypi_0
pip=24.0=pyhd8ed1ab_0
py4j=0.10.9.7=pypi_0
pymongo=4.8.0=pypi_0
pyspark=3.5.1=pypi_0
python=3.9.19=hd7ebdb9_0_cpython
python-dateutil=2.9.0.post0=pypi_0
pytz=2024.1=pypi_0
readline=8.2=h92ec313_1
setuptools=71.0.4=pyhd8ed1ab_0
six=1.16.0=pypi_0
tk=8.6.13=h5083fa2_1
tzdata=2024.1=pypi_0
wheel=0.43.0=pyhd8ed1ab_1
xz=5.2.6=h57fd34a_0
